+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,59 @@
 # @orderly/docs
 
+## 0.3.2
+
+### Patch Changes
+
+- fix bugs;
+- Updated dependencies
+  - @orderly.network/react@1.3.2
+  - @orderly.network/web3-onboard@1.3.2
+  - @orderly.network/hooks@1.3.2
+  - @orderly.network/types@1.3.2
+
+## 0.3.1
+
+### Patch Changes
+
+- fix bugs;
+- Updated dependencies
+  - @orderly.network/react@1.3.1
+  - @orderly.network/web3-onboard@1.3.1
+  - @orderly.network/hooks@1.3.1
+  - @orderly.network/types@1.3.1
+
+## 0.3.0
+
+### Minor Changes
+
+- fd452911: update;
+
+### Patch Changes
+
+- e4edd4c7: publish referral
+- fix bugs;
+- Updated dependencies [e4edd4c7]
+- Updated dependencies
+- Updated dependencies [fd452911]
+  - @orderly.network/react@1.3.0
+  - @orderly.network/web3-onboard@1.3.0
+  - @orderly.network/hooks@1.3.0
+  - @orderly.network/types@1.3.0
+
+## 0.3.0-alpha.0
+
+### Minor Changes
+
+- update;
+
+### Patch Changes
+
+- Updated dependencies
+  - @orderly.network/react@1.3.0-alpha.0
+  - @orderly.network/web3-onboard@1.3.0-alpha.0
+  - @orderly.network/hooks@1.3.0-alpha.0
+  - @orderly.network/types@1.3.0-alpha.0
+
 ## 0.2.10
 
 ### Patch Changes