+++ packages/create-orderly-app/CHANGELOG.md
@@ -1,5 +1,34 @@
 # create-orderly-app
 
+## 0.3.2
+
+### Patch Changes
+
+- fix bugs;
+
+## 0.3.1
+
+### Patch Changes
+
+- fix bugs;
+
+## 0.3.0
+
+### Minor Changes
+
+- fd452911: update;
+
+### Patch Changes
+
+- e4edd4c7: publish referral
+- fix bugs;
+
+## 0.3.0-alpha.0
+
+### Minor Changes
+
+- update;
+
 ## 0.2.8
 
 ### Patch Changes